moving intial model creation from .api to .tune
